refactor: firebase module
