+++ src/modules/messages/firebase/firebase.const.ts
@@ -1,0 +1,3 @@
+export const FIREBASE_INITIALIZATION_OPTION = Symbol.for(
+  "FIREBASE_INITIALIZATION_OPTION"
+);
